added calls to updateValidatorData to validationFails() and validationPasses() to fix a validation bug.
The validator is initialized in the constructor where the data
(attributes) array is passed to its constructor upon instantiation. If
the BaseEloquent class is instantiated without data an empty data set
would be provided to the validator. When the validation function is
called the validator data would not be updated, and an error message
would result. Updating the validator's data prevents this from occuring.